moved the useHeight function to the API folder.
also imported it into expandable container
